requirements.txt
flask
flask-cors
python-dotenv
langchain
langchain-google-genai
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-dateutil
gunicorn
